+++ src/app/HighLevelAPI.hs
@@ -7,7 +7,7 @@ import Data.Geolocation.GEOS
 -- which guarantees that they are released when the context goes out of scope
 demo :: IO ()
 demo = do
-    withContext $ \ctx -> do
+    withGEOS $ \ctx -> do
         reader <- mkReader ctx
         g0 <- readGeometry reader "POLYGON (( 10 10, 10 20, 20 20, 20 10, 10 10 ))"
         g1 <- readGeometry reader "POLYGON (( 11 11, 11 12, 12 12, 12 11, 11 11 ))"